Title: 소수점 반올림하기, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240102/소수점 반올림하기/rounding-decimal-points.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-       int a = 25.352;
+       double a = 25.352;
        System.ouut.printf("%1f",a); // 여기에 코드를 작성해주세요.
     }
 }
